Update load_row/col functions to use 8bit coordinates

+++ include/Scroll.h
@@ -20,8 +20,8 @@ extern INT16 draw_scroll_x;
 extern INT16 draw_scroll_y;
 extern UINT16 scroll_x_max;
 extern UINT16 scroll_y_max;
-extern INT16 scroll_offset_x;
-extern INT16 scroll_offset_y;
+extern BYTE scroll_offset_x;
+extern BYTE scroll_offset_y;
 extern UINT8 pending_w_i;
 extern UINT8 pending_h_i;
